// EasyBuy-Frontend-Customer/src/app/models/user.model.ts
export interface User {
    id?: string;
    name: string;
    email: string;
    phone: string;
    password: string;
    address: string;
    status: string;
    role: string;
}
export interface Users {
    id?: number;
    name: string;
    email: string;
    phone: string;
    password: string;
    address: string;
    status: number;
    role: number;
}